services/balance-orchestrator/backend/pyproject.toml
[project]
name = "balance-plus"
version = "0.1.0"
description = ""
authors = [
    {name = "lrshlyogin",email = "ooooooorrrrr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-gitlab (>=7.0.0,<8.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "python-slugify (>=8.0.4,<9.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"